# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kwaaka Integration API",
        "title": "Integration API",
        "contact": {
            "name": "Kwaaka",
            "email": "neversi123123@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "staging.kwaaka.com",
    "basePath": "/v1",
    "paths": {
...
        "/jowi/events": {
            "post": {
                "tags": [
                    "jowi"
                ],
                "summary": "events",
                "parameters": [
                    {
                        "description": "event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Data": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "count": {
                    "type": "string"
                },
                "courier_name": {
                    "type": "string"
                },
                "courier_phone": {
                    "type": "string"
                },
                "device_id": {
                    "description": "stoplist fields",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "order_id": {
                    "description": "order fields",
                    "type": "string"
                },
                "restaurant_id": {
                    "description": "integration_api fields",
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Event": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Data"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"Token from cognito\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}